requirements.txt
pytesseract
Pillow
watchdog
